+++ turbopack/crates/turbo-tasks/src/lib.rs
@@ -56,11 +56,11 @@ mod invalidation;
 mod join_iter_ext;
 mod key_value_pair;
 pub mod keyed;
-pub mod keyed_read_ref;
 #[doc(hidden)]
 pub mod macro_helpers;
 mod magic_any;
 mod manager;
+pub mod mapped_read_ref;
 mod marker_trait;
 pub mod message_queue;
 mod native_function;
@@ -119,6 +119,7 @@ pub use crate::{
         mark_session_dependent, prevent_gc, run, run_once, run_once_with_reason, trait_call,
         turbo_tasks, turbo_tasks_scope, turbo_tasks_weak, with_turbo_tasks,
     },
+    mapped_read_ref::MappedReadRef,
     output::OutputContent,
     raw_vc::{CellId, RawVc, ReadRawVcFuture, ResolveTypeError},
     read_options::{ReadCellOptions, ReadOutputOptions},
